+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## master
 
-- CHANGED: Deprecated Dnsimple::Client.base_uri= in favor of Dnsimple::Client.host=.
+- CHANGED: Deprecated Dnsimple::Client.base_uri= in favor of Dnsimple::Client.site=.
 
 - CHANGED: Normalized exception handling. No more RuntimeError.
   In case of request error, the client raises RequestError, RecordExists or RecodNotFound